requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
azure-ai-vision-imageanalysis==1.0.0b1
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
